Added file for task 2

+++ 0x18-dynamic_libraries/100-operations.c
@@ -1,11 +1,9 @@
 #include <stdio.h>
-#include <stdlib.h>
-#include <stddef.h>
 
 int add(int a, int b);
 int sub(int a, int b);
 int mul(int a, int b);
-int div(int a, int b);
+float div(int a, int b);
 int mod(int a, int b);
 
 /**
@@ -54,7 +52,7 @@ int mul(int a, int b)
  * Return: a / b
  */
 
-int div(int a, int b)
+float div(int a, int b)
 {
         return (a / b);
 }
